Tranporting lower level error (#26)
* Using original io::Error

Transport lower level error for more transparency.

* Shall we use the original IntegrationError?

Would be useful to get more information why did the integration failed?
I believe that we don't have much how to address the possible errors
here, is that right?

* Stop warning for unused Error::Undefined

Let's allow dead_code for Undefined error while we are on development
mode. Once we get a stable library we reconsider if we actually need
this type or error.