refactor: ConstantDepth (#30)
* syntax: Running 'cargo fmt'

* Moving ConstantDepth into bathymetry module

* fix: Loading ConstantDepth from bathymetry

* feat: Builder for ConstantDepth

* test: ConstantDepth with default h

* refactor: Organizing module bathymetry into a directory

* refactor: Organizing ConstantDepth in a file

It seems like there will be a few support material for each type of
bathymetry, such as specific tests and documentation. Let's isolate each
type in one file to keep the code cleaner and easier to read.

* feat: ConstantDepth::builder()

* Tagging x & y as not useful

ConstantDepth doesn't care about the location, but the depth is always
the same.

* Respect the general behavior of NaN inputs

* Allow dead code for now

Let's stop warnings to make it easier to keep track on suggestions and
quality of the code. We might later turn into some level of public, but
let's keep it private until we know what we need.