feat: New type of bathymetry - constant slope
It has an infinite domain with given constant slope (dx, dy), and
reference depth (h0) for a given location (x0, y0).

This is another application for idealized scenarios. With it we can
avoid creating a bathymetry file as well as avoid I/O. This will be
particularly useful for tests, reducing the computational cost.