Minor paper edits (#241)
* Fix typo in the python examples

This commit fixes issue #239. Some parts of the code example shown in the paper used phi0 as the initial direction, when our convention is theta0.

* Revise acknowledgements in paper.md

Updated acknowledgements section to include additional support and reviewers. Fixes issue #231

* Updates dispersion relationship

Updates the convention in the dispersion relationship equation to comply with NREL's internal review.